Replace Double with BigDecimal

+++ src/main/java/com/rocketcredit/user_data/model/Transaction.java
@@ -3,6 +3,7 @@ package com.rocketcredit.user_data.model;
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalDate;
+import java.math.BigDecimal;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import jakarta.validation.Valid;
@@ -27,7 +28,7 @@ public class Transaction {
   @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
   private LocalDate date;
 
-  private Double amount;
+  private BigDecimal amount;
 
   private String currency;
 
@@ -113,7 +114,7 @@ public class Transaction {
     this.date = date;
   }
 
-  public Transaction amount(Double amount) {
+  public Transaction amount(BigDecimal amount) {
     this.amount = amount;
     return this;
   }
@@ -125,11 +126,11 @@ public class Transaction {
   
   @Schema(name = "amount", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
   @JsonProperty("amount")
-  public Double getAmount() {
+  public BigDecimal getAmount() {
     return amount;
   }
 
-  public void setAmount(Double amount) {
+  public void setAmount(BigDecimal amount) {
     this.amount = amount;
   }
 
@@ -200,4 +201,3 @@ public class Transaction {
     return o.toString().replace("\n", "\n    ");
   }
 }
-
